requirements.txt
asgiref==3.11.1
beautifulsoup4==4.13.4
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cssselect2==0.9.0
cssutils==2.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.62.1
idna==3.11
more-itertools==10.7.0
pillow==12.1.1
pycparser==3.0
pydyf==0.12.1
PyJWT==2.10.1
PyPDF2==3.0.1
pyphen==0.17.2
python-decouple==3.8
requests==2.33.1
soupsieve==2.7
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.1.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
